test: extract host daemon e2e helpers
