+++ test/e2e/docker_jupyter_host_daemon_test.go
@@ -1,19 +1,15 @@
 package e2e
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
-	"net"
 	"net/http"
 	"net/url"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strconv"
 	"strings"
-	"sync"
 	"testing"
 	"time"
 
@@ -45,7 +41,6 @@ func TestE2EDockerJupyterHostDaemonStopResume(t *testing.T) {
 	}
 	t.Cleanup(func() { _ = os.RemoveAll(testRoot) })
 	dockerClient := newE2EDockerClient(t, ctx, image)
-	t.Cleanup(func() { _ = dockerClient.Close() })
 	if hostname, hostErr := os.Hostname(); hostErr == nil {
 		if _, inspectErr := dockerClient.ContainerInspect(ctx, hostname); inspectErr == nil {
 			t.Skip("host-daemon Jupyter E2E requires the test runner to run outside Docker")
@@ -56,17 +51,10 @@ func TestE2EDockerJupyterHostDaemonStopResume(t *testing.T) {
 	listenAddress := unusedLoopbackAddress(t)
 	baseURL := "http://" + listenAddress
 	daemon := startE2EDaemon(t, binary, repoRoot, testRoot, listenAddress, image)
-	t.Cleanup(func() { daemon.stop(t) })
 	waitForE2EDaemon(t, ctx, daemon, baseURL)
 
-	httpClient := &http.Client{
-		Timeout: 3 * time.Minute,
-		Transport: func() *http.Transport {
-			transport := http.DefaultTransport.(*http.Transport).Clone()
-			transport.Proxy = nil
-			return transport
-		}(),
-	}
+	httpClient := newE2EHTTPClient()
+	httpClient.Timeout = 3 * time.Minute
 	projectClient := agentcomposev2connect.NewProjectServiceClient(httpClient, baseURL)
 	runClient := agentcomposev2connect.NewRunServiceClient(httpClient, baseURL)
 	sandboxClient := agentcomposev2connect.NewSandboxServiceClient(httpClient, baseURL)
@@ -163,138 +151,6 @@ func TestE2EDockerJupyterHostDaemonStopResume(t *testing.T) {
 	t.Logf("Docker Jupyter binding refreshed from persisted stale port to 127.0.0.1:%d", resumedState.HostPort)
 }
 
-type synchronizedBuffer struct {
-	mu sync.Mutex
-	b  bytes.Buffer
-}
-
-func (b *synchronizedBuffer) Write(data []byte) (int, error) {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-	return b.b.Write(data)
-}
-
-func (b *synchronizedBuffer) String() string {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-	return b.b.String()
-}
-
-type e2eDaemonProcess struct {
-	cmd  *exec.Cmd
-	done chan error
-	logs synchronizedBuffer
-}
-
-func startE2EDaemon(t *testing.T, binary, repoRoot, testRoot, listenAddress, image string) *e2eDaemonProcess {
-	t.Helper()
-	process := &e2eDaemonProcess{done: make(chan error, 1)}
-	process.cmd = exec.Command(binary, "daemon")
-	process.cmd.Dir = repoRoot
-	process.cmd.Env = overrideE2EEnv(os.Environ(), map[string]string{
-		"AGENT_COMPOSE_SOCKET":     filepath.Join(testRoot, "agent-compose.sock"),
-		"AUTH_PASSWORD":            "",
-		"AUTH_USERNAME":            "",
-		"DATA_ROOT":                testRoot,
-		"DEFAULT_IMAGE":            image,
-		"DOCKER_DEFAULT_IMAGE":     image,
-		"DOCKER_HOST_SANDBOX_ROOT": filepath.Join(testRoot, "sandboxes"),
-		"DOCKER_HOST_SESSION_ROOT": "",
-		"HTTP_BASIC_AUTH":          "",
-		"HTTP_LISTEN":              listenAddress,
-		"JUPYTER_PROXY_BASE":       "/jupyter",
-		"JUPYTER_READY_TIMEOUT":    "2m",
-		"LLM_API_ENDPOINT":         "",
-		"LLM_API_KEY":              "",
-		"OPENAI_API_KEY":           "",
-		"RUNTIME_DRIVER":           "docker",
-		"SANDBOX_ROOT":             filepath.Join(testRoot, "sandboxes"),
-		"SANDBOX_START_TIMEOUT":    "3m",
-	})
-	process.cmd.Stdout = &process.logs
-	process.cmd.Stderr = &process.logs
-	if err := process.cmd.Start(); err != nil {
-		t.Fatalf("start agent-compose daemon: %v", err)
-	}
-	go func() { process.done <- process.cmd.Wait() }()
-	return process
-}
-
-func (p *e2eDaemonProcess) stop(t *testing.T) {
-	t.Helper()
-	if p == nil || p.cmd == nil || p.cmd.Process == nil {
-		return
-	}
-	_ = p.cmd.Process.Signal(os.Interrupt)
-	select {
-	case err := <-p.done:
-		if err != nil {
-			t.Logf("agent-compose daemon exit: %v", err)
-		}
-	case <-time.After(15 * time.Second):
-		_ = p.cmd.Process.Kill()
-		<-p.done
-		t.Log("agent-compose daemon required forced termination")
-	}
-}
-
-func waitForE2EDaemon(t *testing.T, ctx context.Context, daemon *e2eDaemonProcess, baseURL string) {
-	t.Helper()
-	transport := http.DefaultTransport.(*http.Transport).Clone()
-	transport.Proxy = nil
-	client := &http.Client{Timeout: time.Second, Transport: transport}
-	deadline := time.Now().Add(30 * time.Second)
-	for time.Now().Before(deadline) {
-		req, err := http.NewRequestWithContext(ctx, http.MethodGet, baseURL+"/api/version", nil)
-		if err == nil {
-			resp, requestErr := client.Do(req)
-			if requestErr == nil {
-				_ = resp.Body.Close()
-				if resp.StatusCode == http.StatusOK {
-					return
-				}
-			}
-		}
-		time.Sleep(100 * time.Millisecond)
-	}
-	t.Fatalf("agent-compose daemon did not become ready at %s\ndaemon log:\n%s", baseURL, daemon.logs.String())
-}
-
-func e2eDaemonBinary(t *testing.T, ctx context.Context, repoRoot, testRoot string) string {
-	t.Helper()
-	if configured := strings.TrimSpace(os.Getenv("AGENT_COMPOSE_E2E_BINARY")); configured != "" {
-		binary, err := filepath.Abs(configured)
-		if err != nil {
-			t.Fatalf("resolve AGENT_COMPOSE_E2E_BINARY: %v", err)
-		}
-		return binary
-	}
-	binary := filepath.Join(testRoot, "agent-compose")
-	cmd := exec.CommandContext(ctx, "go", "build", "-o", binary, "./cmd/agent-compose")
-	cmd.Dir = repoRoot
-	if output, err := cmd.CombinedOutput(); err != nil {
-		t.Fatalf("build agent-compose daemon: %v\n%s", err, output)
-	}
-	return binary
-}
-
-func newE2EDockerClient(t *testing.T, ctx context.Context, image string) *client.Client {
-	t.Helper()
-	dockerClient, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
-	if err != nil {
-		t.Fatalf("create Docker client: %v", err)
-	}
-	if _, err := dockerClient.Ping(ctx); err != nil {
-		_ = dockerClient.Close()
-		t.Fatalf("Docker daemon is required: %v", err)
-	}
-	if _, err := dockerClient.ImageInspect(ctx, image); err != nil {
-		_ = dockerClient.Close()
-		t.Fatalf("Docker Jupyter image %q is required: %v", image, err)
-	}
-	return dockerClient
-}
-
 func inspectE2EDockerJupyterPort(t *testing.T, ctx context.Context, dockerClient *client.Client, sandboxID string, guestPort int) (string, int) {
 	t.Helper()
 	args := filters.NewArgs(
@@ -385,70 +241,6 @@ func writeE2EProxyState(t *testing.T, path string, state domain.ProxyState) {
 	}
 	if err := os.WriteFile(path, append(data, '\n'), 0o644); err != nil {
 		t.Fatalf("write proxy state: %v", err)
-	}
-}
-
-func removeE2EDockerSandboxFallback(t *testing.T, ctx context.Context, dockerClient *client.Client, sandboxID string) {
-	t.Helper()
-	args := filters.NewArgs(filters.Arg("label", "agent-compose.sandbox_id="+sandboxID))
-	containers, err := dockerClient.ContainerList(ctx, containerapi.ListOptions{All: true, Filters: args})
-	if err != nil {
-		t.Logf("fallback Docker sandbox lookup failed: %v", err)
-		return
-	}
-	for _, item := range containers {
-		if err := dockerClient.ContainerRemove(ctx, item.ID, containerapi.RemoveOptions{Force: true}); err != nil {
-			t.Logf("fallback Docker sandbox removal failed for %s: %v", item.ID, err)
-		}
-	}
-}
-
-func unusedLoopbackAddress(t *testing.T) string {
-	t.Helper()
-	listener, err := net.Listen("tcp", "127.0.0.1:0")
-	if err != nil {
-		t.Fatalf("allocate daemon listen address: %v", err)
-	}
-	address := listener.Addr().String()
-	if err := listener.Close(); err != nil {
-		t.Fatalf("release daemon listen address: %v", err)
-	}
-	return address
-}
-
-func overrideE2EEnv(environ []string, overrides map[string]string) []string {
-	values := make(map[string]string, len(environ)+len(overrides))
-	for _, item := range environ {
-		name, value, ok := strings.Cut(item, "=")
-		if ok {
-			values[name] = value
-		}
-	}
-	for name, value := range overrides {
-		values[name] = value
-	}
-	result := make([]string, 0, len(values))
-	for name, value := range values {
-		result = append(result, name+"="+value)
-	}
-	return result
-}
-
-func e2eRepoRoot(t *testing.T) string {
-	t.Helper()
-	dir, err := os.Getwd()
-	if err != nil {
-		t.Fatalf("get working directory: %v", err)
-	}
-	for {
-		if _, err := os.Stat(filepath.Join(dir, "go.mod")); err == nil {
-			return dir
-		}
-		parent := filepath.Dir(dir)
-		if parent == dir {
-			t.Fatalf("find repository root from %s", dir)
-		}
-		dir = parent
 	}
 }
